[FE-217] Refactoring ButtonTx and Deposit.
