+++ src/components/UserActions/Deposit/Deposit.jsx
@@ -1,5 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Form, Input, Dropdown } from 'semantic-ui-react';
+import { web3FromAddress } from '@polkadot/extension-dapp';
+import { connect } from 'react-redux';
 
 import { UNDERLYING_ASSETS_TYPES } from '../../../util/constants';
 
@@ -8,7 +10,14 @@ import ButtonTx from '../../../util/ButtonTx';
 
 import classes from './Deposit.module.css';
 // TODO refactoring
-function Deposit({ account, setStateStale, stateStale, updateData }) {
+function Deposit({
+	account,
+	setStateStale,
+	stateStale,
+	updateData,
+	api,
+	keyring,
+}) {
 	const [amount, setAmount] = useState(0);
 	const [asset, setAsset] = useState('');
 	const [loading, setLoading] = useState(false);
@@ -42,6 +51,66 @@ function Deposit({ account, setStateStale, stateStale, updateData }) {
 		return <Loading />;
 	}
 
+	const updateContentPool = () => {
+		if (typeof updateData === 'function') {
+			updateData();
+		}
+	};
+
+	const palletName = 'minterestProtocol';
+	const transactionName = 'depositUnderlying';
+
+	const sendTransaction = async () => {
+		setLoading(true);
+		const currentUser = keyring.getPair(account);
+		try {
+			if (currentUser.isLocked) {
+				const injector = await web3FromAddress(account);
+				await api.tx[palletName][transactionName](asset, amount).signAndSend(
+					account,
+					{ signer: injector.signer },
+					transactionCallback
+				);
+			} else {
+				await api.tx[palletName][transactionName](asset, amount).signAndSend(
+					currentUser,
+					transactionCallback
+				);
+			}
+		} catch (err) {
+			alert(err.toString());
+			setLoading(false);
+			setStateStale(!stateStale);
+		}
+
+		setInitialStates();
+	};
+
+	const transactionCallback = ({ events = [], status }) => {
+		if (status.isFinalized) {
+			setLoading(false);
+			events.forEach(
+				({
+					event: {
+						method,
+						section,
+						data: [error],
+					},
+				}) => {
+					if (section === 'system' && method === 'ExtrinsicSuccess') {
+						updateContentPool();
+						alert('Transaction completed successfully.');
+					} else if (method === 'ExtrinsicFailed' && error.isModule) {
+						const decoded = api.registry.findMetaError(error.asModule);
+						const { documentation } = decoded;
+						alert(`${documentation.join(' ')}`);
+					}
+				}
+			);
+			setStateStale(!stateStale);
+		}
+	};
+
 	return (
 		<Form className={classes.deposit}>
 			<Input
@@ -59,20 +128,20 @@ function Deposit({ account, setStateStale, stateStale, updateData }) {
 				onChange={onChangeAsset}
 			/>
 			<ButtonTx
-				account={account}
-				transactionParams={[asset, amount]}
-				setStateStale={setStateStale}
-				stateStale={stateStale}
-				setLoading={setLoading}
 				isInvalid={isInvalid}
 				setInitialStates={setInitialStates}
 				buttonLabel={'Deposit'}
-				palletName={'minterestProtocol'}
-				transactionName={'depositUnderlying'}
 				updateData={updateData}
+				onClick={sendTransaction}
+				color={account ? 'green' : 'red'}
 			/>
 		</Form>
 	);
 }
 
-export default Deposit;
+const mapStateToProps = (state) => ({
+	api: state.substrate.api,
+	keyring: state.account.keyring,
+});
+
+export default connect(mapStateToProps, null)(Deposit);
